refactor(src/): people, ladder, items 패키지 생성 및 파일 위치 변경
- Height, Count 상수 추출
