+++ src/main/java/controller/LadderGame.java
@@ -3,11 +3,11 @@ package controller;
 import dto.LadderInfo;
 import dto.ResultInfo;
 import java.util.List;
-import model.Height;
-import model.Items;
-import model.Ladder;
-import model.People;
-import model.PersonCount;
+import model.ladder.Height;
+import model.items.Items;
+import model.ladder.Ladder;
+import model.people.People;
+import model.people.PersonCount;
 import model.Result;
 import model.line.RandomLinesGenerator;
 import view.InputView;
