+++ src/main/java/dto/LadderInfo.java
@@ -1,9 +1,9 @@
 package dto;
 
 import java.util.List;
-import model.Items;
-import model.Ladder;
-import model.People;
+import model.items.Items;
+import model.ladder.Ladder;
+import model.people.People;
 
 public record LadderInfo(List<String> peopleNames, List<LineInfo> lines, List<String> itemNames) {
     public static LadderInfo from(final People people, final Ladder ladder, final Items items) {
