+++ src/main/java/model/items/Item.java
@@ -1,10 +1,12 @@
-package model;
+package model.items;
+
+import model.Name;
 
 public class Item {
     private final Name name;
 
-    public Item(final String name) {
-        this.name = new ItemName(name);
+    public Item(final String rawName) {
+        this.name = new ItemName(rawName);
     }
 
     public String getName() {
