+++ src/main/java/model/ladder/Height.java
@@ -1,6 +1,8 @@
-package model;
+package model.ladder;
 
 public class Height {
+    private static final int MINIMUM_HEIGHT = 1;
+
     public final int height;
 
     public Height(final int height) {
@@ -9,7 +11,7 @@ public class Height {
     }
 
     private void validateHeight(final int height) {
-        if (height < 1) {
+        if (height < MINIMUM_HEIGHT) {
             throw new IllegalArgumentException("사다리의 높이는 1 이상이어야 합니다.");
         }
     }
