+++ src/main/java/model/line/LinesGenerator.java
@@ -1,7 +1,7 @@
 package model.line;
 
 import java.util.List;
-import model.Height;
+import model.ladder.Height;
 
 public interface LinesGenerator {
     List<Line> generate(final Height height, final int pathCount);
