+++ src/test/java/model/ResultTest.java
@@ -3,6 +3,8 @@ package model;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Map;
+import model.items.Item;
+import model.people.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
