+++ src/test/java/model/items/ItemNameTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.items;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
