+++ src/test/java/model/items/ItemTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.items;
 
 import static org.assertj.core.api.Assertions.assertThat;
