+++ src/test/java/model/ladder/LadderTest.java
@@ -1,12 +1,13 @@
-package model;
+package model.ladder;
 
-import static model.Path.EXIST;
-import static model.Path.NOT_EXIST;
+import static model.line.Path.EXIST;
+import static model.line.Path.NOT_EXIST;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import model.line.Line;
 import model.line.FixedLinesGenerator;
+import model.people.PersonCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
