+++ src/test/java/model/line/LineTest.java
@@ -1,9 +1,8 @@
-package model;
+package model.line;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import model.line.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
