+++ src/test/java/model/people/PersonCountTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.people;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
